Refactored (java-application): Splitted Config-Module into interface and implementation
