+++ server/src/main/java/Config/Impl/ConfigJson.java
@@ -1,6 +1,6 @@
-package Config;
+package Config.Impl;
 
-import graphql.validation.rules.KnownTypeNames;
+import Config.Config;
 import lombok.Getter;
 
 import java.util.Map;
