+++ server/src/main/java/Config/Impl/JsonValideException.java
@@ -1,4 +1,6 @@
-package Config;
+package Config.Impl;
+
+import Config.Config;
 
 public class JsonValideException extends Exception
 {
